AD-288 refactor packages and code style
