+++ analytics/src/commonMain/kotlin/dev/icerock/moko/analytics/library/AnalyticsAdapter.kt
@@ -2,9 +2,9 @@
  * Copyright 2021 IceRock MAG Inc. Use of this source code is governed by the Apache 2.0 license.
  */
 
-package dev.icerock.moko.analytics.library.analytics.presentation
+package dev.icerock.moko.analytics.library
 
-import dev.icerock.moko.analytics.library.analytics.model.AnalyticsEvent
+import dev.icerock.moko.analytics.library.AnalyticsEvent
 
 interface AnalyticsAdapter {
     fun sendEvent(analyticsEvent: AnalyticsEvent)
