+++ analytics/src/commonMain/kotlin/dev/icerock/moko/analytics/library/AnalyticsEvent.kt
@@ -1,7 +1,7 @@
 /*
  * Copyright 2021 IceRock MAG Inc. Use of this source code is governed by the Apache 2.0 license.
  */
-package dev.icerock.moko.analytics.library.analytics.model
+package dev.icerock.moko.analytics.library
 
 interface AnalyticsEvent
