+++ analytics/src/commonMain/kotlin/dev/icerock/moko/analytics/library/AnalyticsTracker.kt
@@ -1,10 +1,10 @@
 /*
  * Copyright 2021 IceRock MAG Inc. Use of this source code is governed by the Apache 2.0 license.
  */
-package dev.icerock.moko.analytics.library.analytics.di
+package dev.icerock.moko.analytics.library
 
-import dev.icerock.moko.analytics.library.analytics.model.AnalyticsEvent
-import dev.icerock.moko.analytics.library.analytics.presentation.AnalyticsAdapter
+import dev.icerock.moko.analytics.library.AnalyticsEvent
+import dev.icerock.moko.analytics.library.AnalyticsAdapter
 import io.github.aakira.napier.Napier
 
 class AnalyticsTracker(private val adapterAmplitudes: List<AnalyticsAdapter>) {
